Move internal/rule to rule
