+++ cmd/gazelle/BUILD.bazel
@@ -23,10 +23,10 @@ go_library(
         "//internal/merger:go_default_library",
         "//internal/repos:go_default_library",
         "//internal/resolve:go_default_library",
-        "//internal/rule:go_default_library",
         "//internal/version:go_default_library",
         "//internal/walk:go_default_library",
         "//label:go_default_library",
+        "//rule:go_default_library",
         "//vendor/github.com/pmezard/go-difflib/difflib:go_default_library",
     ],
 )
