+++ cmd/gazelle/print.go
@@ -19,7 +19,7 @@ import (
 	"os"
 
 	"github.com/bazelbuild/bazel-gazelle/internal/config"
-	"github.com/bazelbuild/bazel-gazelle/internal/rule"
+	"github.com/bazelbuild/bazel-gazelle/rule"
 )
 
 func printFile(c *config.Config, f *rule.File) error {
