+++ internal/config/BUILD.bazel
@@ -9,8 +9,8 @@ go_library(
     importpath = "github.com/bazelbuild/bazel-gazelle/internal/config",
     visibility = ["//visibility:public"],
     deps = [
-        "//internal/rule:go_default_library",
         "//internal/wspace:go_default_library",
+        "//rule:go_default_library",
     ],
 )
 
@@ -19,5 +19,5 @@ go_test(
     size = "small",
     srcs = ["config_test.go"],
     embed = [":go_default_library"],
-    deps = ["//internal/rule:go_default_library"],
+    deps = ["//rule:go_default_library"],
 )
