+++ internal/language/go/BUILD.bazel
@@ -52,9 +52,9 @@ go_library(
         "//internal/language/proto:go_default_library",
         "//internal/repos:go_default_library",
         "//internal/resolve:go_default_library",
-        "//internal/rule:go_default_library",
         "//label:go_default_library",
         "//pathtools:go_default_library",
+        "//rule:go_default_library",
         "//vendor/github.com/bazelbuild/buildtools/build:go_default_library",
     ],
 )
@@ -78,10 +78,10 @@ go_test(
         "//internal/merger:go_default_library",
         "//internal/repos:go_default_library",
         "//internal/resolve:go_default_library",
-        "//internal/rule:go_default_library",
         "//internal/testtools:go_default_library",
         "//internal/walk:go_default_library",
         "//label:go_default_library",
+        "//rule:go_default_library",
         "//vendor/github.com/bazelbuild/buildtools/build:go_default_library",
         "//vendor/golang.org/x/tools/go/vcs:go_default_library",
     ],
