+++ internal/merger/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
     ],
     importpath = "github.com/bazelbuild/bazel-gazelle/internal/merger",
     visibility = ["//visibility:public"],
-    deps = ["//internal/rule:go_default_library"],
+    deps = ["//rule:go_default_library"],
 )
 
 go_test(
@@ -20,6 +20,6 @@ go_test(
         "//internal/language:go_default_library",
         "//internal/language/go:go_default_library",
         "//internal/language/proto:go_default_library",
-        "//internal/rule:go_default_library",
+        "//rule:go_default_library",
     ],
 )
