+++ internal/repos/BUILD.bazel
@@ -11,9 +11,9 @@ go_library(
     importpath = "github.com/bazelbuild/bazel-gazelle/internal/repos",
     visibility = ["//visibility:public"],
     deps = [
-        "//internal/rule:go_default_library",
         "//label:go_default_library",
         "//pathtools:go_default_library",
+        "//rule:go_default_library",
         "//vendor/github.com/pelletier/go-toml:go_default_library",
         "//vendor/golang.org/x/tools/go/vcs:go_default_library",
     ],
@@ -29,7 +29,7 @@ go_test(
     ],
     embed = [":go_default_library"],
     deps = [
-        "//internal/rule:go_default_library",
+        "//rule:go_default_library",
         "//vendor/golang.org/x/tools/go/vcs:go_default_library",
     ],
 )
