+++ internal/resolve/BUILD.bazel
@@ -11,7 +11,7 @@ go_library(
     deps = [
         "//internal/config:go_default_library",
         "//internal/repos:go_default_library",
-        "//internal/rule:go_default_library",
         "//label:go_default_library",
+        "//rule:go_default_library",
     ],
 )
