+++ rule/directives.go
