+++ rule/directives_test.go
