+++ rule/merge.go
