+++ rule/platform.go
