+++ rule/platform_strings.go
