+++ rule/rule.go
