+++ rule/rule_test.go
